+++ CHANGES.rst
@@ -1,5 +1,5 @@
-unreleased
-==========
+2.0 (2022-01-02)
+================
 
 - Drop support for Python 2.7, 3.4, 3.5, 3.6.
 
@@ -34,6 +34,8 @@ unreleased
 
 - Remove the dependency on ``setuptools`` for resolving pkg_resources-style
   importable paths in ``colander.GlobalObject``.
+
+- Refresh localization files with Babel 2.11.
 
 1.8.3 (2020-11-28)
 ==================